+++ apps/server/migrations/20260321113353_add_supply_entry.js
@@ -1,0 +1,17 @@
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.up = function (knex) {
+  return knex('supply').update({
+    unreleased_supply: knex.raw('100000000::bigint * 43200000::bigint'),
+  });
+};
+
+/**
+ * @param { import("knex").Knex } knex
+ * @returns { Promise<void> }
+ */
+exports.down = function (knex) {
+  return knex('supply').update({ unreleased_supply: 0 });
+};