+++ packages/core/src/core.ts
@@ -1,8 +1,8 @@
 export abstract class Core {
-  static readonly COIN = 100;
+  static readonly COIN = 100_000_000;
   static amountToString(amount: number, signDisplay: 'always' | 'never' = 'never') {
     return Number(amount).toLocaleString('fi', {
-      maximumFractionDigits: 2,
+      maximumFractionDigits: 8,
       minimumFractionDigits: 2,
       signDisplay,
     });